buildpackage: fix exporting of working copy when run from a subdir
Exporting sources failed when git-buildpackage was run from a
subdirectory of the git working directory because the path of the custom
index file was not handled correctly.

This also fixes a crash in case .git is not a real directory but merely
a normal file containing a reference to the real .git directory (as is
the case with git submodules lately, for example).

Signed-off-by: Markus Lehtonen <markus.lehtonen@linux.intel.com>